server/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
dask==2024.3.1
docstring-inheritance==2.2.0
Flask==3.0.2
fonttools==4.50.0
fsspec==2024.3.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
ndtiff==2.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
partd==1.4.1
pillow==10.2.0
pyasn1==0.5.1
pycromanager==0.29.9
pymmcore==11.1.1.71.0
pyparsing==3.1.2
pyserial==3.5
PySimpleGUI==5.0.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
toolz==0.12.1
urllib3==2.2.1
Werkzeug==3.0.1